requirements.txt
pytest
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest
xlrd
pytest-sugar
pytest-assume
pytest-cover
PyYAML
PyMySQL